Creating new base class called PgpContext
That is the commonalities required to use PGP cryptography, without relying upon GnuPG or local files in any ways, allowing the user to persist his private and public keys as he see fit.
