+++ MimeKit/Cryptography/CryptographyContext.cs
@@ -44,7 +44,7 @@ namespace MimeKit.Cryptography {
 	{
 		const string SubclassAndRegisterFormat = "You need to subclass {0} and then register it with MimeKit.Cryptography.CryptographyContext.Register().";
 		static Func<SecureMimeContext> SecureMimeContextFactory;
-		static Func<OpenPgpContext> OpenPgpContextFactory;
+		static Func<PgpContext> PgpContextFactory;
 		static readonly object mutex = new object ();
 
 		EncryptionAlgorithm[] encryptionAlgorithmRank;
@@ -549,8 +549,8 @@ namespace MimeKit.Cryptography {
 				case "application/pgp-encrypted":
 				case "application/x-pgp-keys":
 				case "application/pgp-keys":
-					if (OpenPgpContextFactory != null)
-						return OpenPgpContextFactory ();
+					if (PgpContextFactory != null)
+						return PgpContextFactory ();
 
 					throw new NotSupportedException (string.Format (SubclassAndRegisterFormat, "MimeKit.Cryptography.OpenPgpContext or MimeKit.Cryptography.GnuPGContext"));
 				default:
@@ -596,9 +596,9 @@ namespace MimeKit.Cryptography {
 				lock (mutex) {
 					SecureMimeContextFactory = () => (SecureMimeContext) ctor.Invoke (new object[0]);
 				}
-			} else if (info.IsSubclassOf (typeof (OpenPgpContext))) {
+			} else if (info.IsSubclassOf (typeof (PgpContext))) {
 				lock (mutex) {
-					OpenPgpContextFactory = () => (OpenPgpContext) ctor.Invoke (new object[0]);
+					PgpContextFactory = () => (PgpContext) ctor.Invoke (new object[0]);
 				}
 			} else {
 				throw new ArgumentException ("The specified type must be a subclass of SecureMimeContext or OpenPgpContext.", nameof (type));
@@ -635,13 +635,13 @@ namespace MimeKit.Cryptography {
 		/// <exception cref="System.ArgumentNullException">
 		/// <paramref name="factory"/> is <c>null</c>.
 		/// </exception>
-		public static void Register (Func<OpenPgpContext> factory) 
+		public static void Register (Func<PgpContext> factory) 
 		{
 			if (factory == null)
 				throw new ArgumentNullException(nameof (factory));
 
 			lock (mutex) {
-				OpenPgpContextFactory = factory;
+				PgpContextFactory = factory;
 			}
 		}
 	}
